这是最后一次提交，现在开始要删减功能，自己封装了一个请求系统图库的类，还有想后台传送广告条的数据的类方法，编写了广告条的展示及数据存储，重新设计了登录界面，就这些吧

+++ 众商智推/Class/Article/View/CityListViewController.m
@@ -162,7 +162,7 @@
 //点击索引回调方法
 - (NSInteger)tableView:(UITableView *)tableView sectionForSectionIndexTitle:(NSString *)title atIndex:(NSInteger)index
 {
-    NSLog(@"title:%@  index:%ld",title,index);
+    NSLog(@"title:%@  index:%ld",title,(long)index);
     return index;
 }
 //标题高
